# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "coco-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/jobs/{id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "operationId": "get_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{id}/events": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "operationId": "job_events",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job event stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CocoErrorKind": {
        "type": "string",
        "description": "High-level error categories for CoCo.\n\nAdding new variants is a breaking change for the public API.",
        "enum": [
          "system",
          "user",
          "network",
          "storage",
          "compute"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Serializable error response used by API layers.",
        "required": [
          "kind",
          "message"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/CocoErrorKind"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
...
      "JobStatusResponse": {
        "type": "object",
        "required": [
          "job_id",
          "status",
          "attempts",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "attempts": {
            "type": "integer",
            "format": "int32"
          },
          "created_at": {
            "type": "string"
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "job_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "version_id": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}